+++ sdk/resourcemanager/storagecache/armstoragecache/skus_client.go
@@ -57,7 +57,6 @@ func NewSKUsClient(subscriptionID string, credential azcore.TokenCredential, opt
 }
 
 // NewListPager - Get the list of StorageCache.Cache SKUs available to this subscription.
-// If the operation fails it returns an *azcore.ResponseError type.
 // Generated from API version 2022-05-01
 // options - SKUsClientListOptions contains the optional parameters for the SKUsClient.List method.
 func (client *SKUsClient) NewListPager(options *SKUsClientListOptions) *runtime.Pager[SKUsClientListResponse] {